Move `load_overridden_dependencies` to own module
This applies following suggestion:
https://github.com/DataDog/datadog-agent/pull/40221#discussion_r2466030991
